pyproject.toml
[tool.poetry]
name = "core-todo"
version = "0.1.0"
description = ""
authors = ["Valerolactone <varanchukv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
config = "^0.5.1"
django-filter = "^24.2"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
redis = "^5.0.7"
pyjwt = "^2.9.0"
requests = "^2.32.3"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.10.1"
pylint = "^3.2.4"
flower = "^2.0.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
